refactor: GameSession으로 게임 인스턴스 관리 로직 분리 및 관리
