+++ src/domain/GameSession.ts
@@ -1,30 +1,65 @@
 import GameSpace from './physics/GameSpace.js';
-import { ScoreDto, SessionStateType } from './game.schema.js';
-import Players from './Players.js';
+import { playerTypeSchema } from './game.schema.js';
+import Ball from './physics/Ball.js';
+import Table from './physics/Table.js';
+import Racket from './physics/Racket.js';
+import { Server } from 'socket.io';
 
 export default class GameSession {
-  private readonly dt = 1 / 60;
+  private readonly gameSpaces: Map<number, GameSpace>;
 
-  constructor(
-    private readonly matchId: number,
-    private readonly engine: GameSpace,
-    private readonly players: Players,
-  ) {}
+  constructor(private readonly io: Server) {
+    this.gameSpaces = new Map<number, GameSpace>();
 
-  tick(): void {
-    // 물리 엔진을 한 스텝 진행합니다.
-    this.engine.step(this.dt);
+    this.startLoop();
   }
 
-  getState(): SessionStateType {
-    return {
-      ball: this.engine.getBallPosition(),
-      racket1: this.engine.getRacket1Position(),
-      racket2: this.engine.getRacket2Position(),
-    };
+  private startLoop() {
+    const fixedTimeStep = 1 / 60; // 60Hz
+    const intervalMs = fixedTimeStep * 1000;
+
+    const timer = setInterval(() => {
+      for (const [matchId, gameSpace] of this.gameSpaces.entries()) {
+        gameSpace.step(fixedTimeStep);
+
+        const message = {
+          ball: gameSpace.getBallPosition(),
+          racket1: gameSpace.getRacket1Position(),
+          racket2: gameSpace.getRacket2Position(),
+        };
+        this.io.to(`match:${matchId}`).emit('game:update', message);
+
+        if (gameSpace.getBallPosition().y <= 0) {
+          clearInterval(timer);
+        }
+      }
+    }, intervalMs);
   }
 
-  getMatchId(): number {
-    return this.matchId;
+  createGameSpace(input: { matchId: number; player1Id: number; player2Id: number }) {
+    if (this.isExist(input.matchId)) {
+      throw new Error(`Game space for match ID ${input.matchId} already exists.`);
+    }
+
+    const ball = new Ball();
+    const table = new Table();
+    const racket1 = new Racket(playerTypeSchema.enum.PLAYER1);
+    const racket2 = new Racket(playerTypeSchema.enum.PLAYER2);
+    const gameSpace = new GameSpace(ball, table, racket1, racket2, input.player1Id);
+
+    this.gameSpaces.set(input.matchId, gameSpace);
+  }
+
+  updateRacketPosition(matchId: number, playerId: number, x: number, y: number, z: number) {
+    const gameSpace = this.gameSpaces.get(matchId);
+    if (!gameSpace) {
+      throw new Error(`Game space for match ID ${matchId} not found.`);
+    }
+
+    gameSpace.updateRacketPosition(playerId, x, y, z);
+  }
+
+  isExist(matchId: number): boolean {
+    return this.gameSpaces.has(matchId);
   }
 }
